Agents/rag/vectara/requirements.txt
chromadb==0.5.5
langchain-text-splitters==0.2.1
python-dotenv==1.0.1
typer==0.12.5
click==8.1.7
pydantic==2.7.4
pypdf==5.1.0
uvicorn==0.29.0
fastapi==0.110.0
modelcontextprotocol==1.0.1
rich==13.7.1
sentence-transformers==2.5.1
